+++ chempy/tests/test_chemistry.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import (absolute_import, division, print_function)
 
-import math
 from operator import attrgetter
 
 import pytest
@@ -10,8 +9,7 @@ from ..util.testing import requires
 from ..util.parsing import parsing_library
 from ..units import default_units, units_library
 from ..chemistry import (
-    Substance, Species, Reaction, ReactionSystem, ArrheniusParam,
-    ArrheniusParamWithUnits, Equilibrium
+    Substance, Species, Reaction, ReactionSystem, Equilibrium
 )
 
 
@@ -134,23 +132,6 @@ def test_ReactionSystem__as_per_substance_array_dict():
         c = rs.as_per_substance_array({'H': 1})
 
     assert rs.as_per_substance_dict([42]) == {'H2O': 42}
-
-
-def test_ArrheniusParam():
-    k = ArrheniusParam(1e10, 42e3)(273.15)
-    ref = 1e10 * math.exp(-42e3/(8.3145*273.15))
-    assert abs((k - ref)/ref) < 1e-4
-
-
-@requires(units_library)
-def test_ArrheniusParamWithUnits():
-    s = default_units.second
-    mol = default_units.mol
-    J = default_units.joule
-    K = default_units.kelvin
-    k = ArrheniusParamWithUnits(1e10/s, 42e3 * J/mol)(273.15*K)
-    ref = 1e10/s * math.exp(-42e3/(8.3145*273.15))
-    assert abs((k - ref)/ref) < 1e-4
 
 
 @requires(units_library)